move format file
